refactoring

+++ src/main/java/io/openslice/model/DeploymentDescriptor.java
@@ -49,7 +49,7 @@ public class DeploymentDescriptor {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private int id = 0;
+	private long id = 0;
 	
 
 	@Basic()
@@ -175,11 +175,11 @@ public class DeploymentDescriptor {
 		this.dateCreated = dateCreated;
 	}
 
-	public int getId() {
+	public long getId() {
 		return id;
 	}
 
-	public void setId(int id) {
+	public void setId(long id) {
 		this.id = id;
 	}
